+++ legend-pure-core/legend-pure-m3-core/src/main/resources/platform/pure/essential/date/creation/date.pure
@@ -65,3 +65,10 @@ function <<PCT.test>> {test.excludePlatform = 'Java compiled'} meta::pure::funct
     assertEquals(%2015-04-16T14:51:59.999, $f->eval(|date(2015, 4, 16, 14, 51, 59.999)));
     assertEquals(%2015-04-16T14:51:59.999, $f->eval(|date(2015, 4, 16, 14, 51, 59.999D)));
 }
+
+function <<PCT.test>> meta::pure::functions::date::tests::testNewDateError<Z|y>(f:Function<{Function<{->Z[y]}>[1]->Z[y]}>[1]):Boolean[1]
+{
+    meta::pure::functions::asserts::assertError(| $f->eval(|date(2016, 13)), 'Invalid month: 13', 71, 61);
+    meta::pure::functions::asserts::assertError(| $f->eval(|date(2016, 12, 32)), 'Invalid day: 2016-12-32', 72, 61);
+    meta::pure::functions::asserts::assertError(| $f->eval(|date(2016, 12, 31, 24)), 'Invalid hour: 24', 73, 61);
+}
