+++ legend-pure-core/legend-pure-m3-core/src/main/resources/platform/pure/essential/date/extract/dayOfMonth.pure
@@ -24,3 +24,8 @@ function <<PCT.test>> meta::pure::functions::date::tests::testDayOfMonth<Z|y>(f:
     assertEquals(15, $f->eval(|%2015-04-15T17:09:21->dayOfMonth()));
     assertEquals(15, $f->eval(|%2015-04-15T17:09:21.398->dayOfMonth()));
 }
+
+function <<PCT.test>> meta::pure::functions::date::tests::testDayOfMonthError<Z|y>(f:Function<{Function<{->Z[y]}>[1]->Z[y]}>[1]):Boolean[1]
+{
+    meta::pure::functions::asserts::assertError(| $f->eval(|%2017->dayOfMonth()), 'Cannot get day of month for 2017', 30, 68);
+}
