--- legend-pure-core/legend-pure-m3-core/src/test/java/org/finos/legend/pure/m3/tests/function/base/date/AbstractTestNewDate.java
@@ -1,29 +1,0 @@
-// Copyright 2025 Goldman Sachs
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//      http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package org.finos.legend.pure.m3.tests.function.base.date;
-
-import org.finos.legend.pure.m3.tests.function.base.PureExpressionTest;
-import org.junit.Test;
-
-public abstract class AbstractTestNewDate extends PureExpressionTest
-{
-    @Test
-    public void testNewDateError() throws Exception
-    {
-        assertExpressionRaisesPureException("Invalid month: 13", 3, 9, "date(2016, 13)");
-        assertExpressionRaisesPureException("Invalid day: 2016-12-32", 3, 9, "date(2016, 12, 32)");
-        assertExpressionRaisesPureException("Invalid hour: 24", 3, 9, "date(2016, 12, 31, 24)");
-    }
-}
