--- legend-pure-core/legend-pure-m3-core/src/test/java/org/finos/legend/pure/m3/tests/function/base/lang/AbstractTestCompare.java
@@ -1,34 +1,0 @@
-// Copyright 2020 Goldman Sachs
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//      http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package org.finos.legend.pure.m3.tests.function.base.lang;
-
-import org.finos.legend.pure.m3.tests.AbstractPureTestWithCoreCompiled;
-import org.junit.Test;
-
-public abstract class AbstractTestCompare extends AbstractPureTestWithCoreCompiled
-{
-    @Test
-    public void testDecimalLongCompare()
-    {
-        compileTestSource(
-                "function test():Boolean[1]\n" +
-                        "{\n" +
-                        " assert(compare(9999999999999992d, 9999999999999991) > 0, |'');\n" +
-                        " assert(compare(9999999999999992d, 9999999999999992) == 0, |'');\n" +
-                        " assert(compare(9999999999999992d, 9999999999999993) < 0, |'');\n" +
-                        "}\n");
-        this.execute("test():Boolean[1]");
-    }
-}
